Move exponential backoff into StreamingTransport
